refactor: extract airport search feature
