+++ src/config/airportSearch.js
@@ -1,0 +1,77 @@
+import { HOME_AIRPORT_COUNTRY } from "./homeAirportDirectory.js";
+
+export const AIRPORT_SEARCH_CONFIG = {
+  country: HOME_AIRPORT_COUNTRY,
+  kind: "all",
+  limit: 12,
+  debounceMs: 220,
+};
+
+export const FEATURED_AIRPORTS = [
+  {
+    icao: "KBOS",
+    iata: "BOS",
+    name: "Boston Logan International Airport",
+    city: "Boston",
+    country: "US",
+    lat: 42.3656,
+    lon: -71.0096,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+  {
+    icao: "KLAX",
+    iata: "LAX",
+    name: "Los Angeles International Airport",
+    city: "Los Angeles",
+    country: "US",
+    lat: 33.9425,
+    lon: -118.4081,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+  {
+    icao: "KJFK",
+    iata: "JFK",
+    name: "John F. Kennedy International Airport",
+    city: "New York",
+    country: "US",
+    lat: 40.6413,
+    lon: -73.7781,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+  {
+    icao: "KORD",
+    iata: "ORD",
+    name: "Chicago O'Hare International Airport",
+    city: "Chicago",
+    country: "US",
+    lat: 41.9742,
+    lon: -87.9073,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+  {
+    icao: "KSFO",
+    iata: "SFO",
+    name: "San Francisco International Airport",
+    city: "San Francisco",
+    country: "US",
+    lat: 37.6213,
+    lon: -122.379,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+  {
+    icao: "KSEA",
+    iata: "SEA",
+    name: "Seattle-Tacoma International Airport",
+    city: "Seattle",
+    country: "US",
+    lat: 47.4502,
+    lon: -122.3088,
+    type: "large_airport",
+    type_label: "Large Airport",
+  },
+];
